Renamed timeseries plot script
